feature/edit-inline-comments
This branch introduces the capability for users to edit their comments inline directly on the post page. When clicking the "Edit" button next to a comment, the displayed text is replaced by a textarea, allowing users to modify their comment. After editing, the "Edit" button changes to "Save", confirming and applying the changes to the comment.