enhancements/saved-posts-ui-footer
Implemented several UI and UX improvements across the application:
Introduced a summary text for the saved posts page, which can be dismissed and remembered using a cookie to enhance the user experience.
Incorporated <Helmet> to manage the title and meta description of the saved posts page for better SEO and user clarity.
Made refinements to the footer section.
Added colors based on their respective categories to the post titles for "Next" and "Previous" navigation options on individual article pages.